Move PartitionsSnap and subclasses into their own module in definitions, import directly in DagsterRun
Summary:
Ensure that we will always be able to deserialize a DagsterRun no matter what type of subset is attached to it.
