+++ python_modules/dagster/dagster/_core/definitions/partitions/snap/__init__.py
@@ -1,0 +1,8 @@
+from dagster._core.definitions.partitions.snap.snap import (
+    DynamicPartitionsSnap as DynamicPartitionsSnap,
+    MultiPartitionsSnap as MultiPartitionsSnap,
+    PartitionDimensionSnap as PartitionDimensionSnap,
+    PartitionsSnap as PartitionsSnap,
+    StaticPartitionsSnap as StaticPartitionsSnap,
+    TimeWindowPartitionsSnap as TimeWindowPartitionsSnap,
+)
