+++ python_modules/dagster/dagster/_core/definitions/partitions/subset/__init__.py
@@ -5,7 +5,7 @@ from dagster._core.definitions.partitions.subset.default import (
     DefaultPartitionsSubset as DefaultPartitionsSubset,
 )
 from dagster._core.definitions.partitions.subset.key_ranges import (
-    KeyRangesPartitionSubset as KeyRangesPartitionSubset,
+    KeyRangesPartitionsSubset as KeyRangesPartitionsSubset,
 )
 from dagster._core.definitions.partitions.subset.partitions_subset import (
     PartitionsSubset as PartitionsSubset,
